+++ src/TsLinkPhp/BaseCL.ts
@@ -1,16 +1,51 @@
 
+export type EventMethodCallData = {
+	name: string,
+	context: any,
+	pars: any[],
+	uploadArgs: number[],
+}
+
+export type EventError = {
+	methodArgs: any[],
+	rawResponse: string|object|null,
+	error: Error|any|null,
+}
+
 export class BaseCL {
 	public context : any = {};
 
 	public url : string = "";
 
-	public onPrepareRequest : ((ev: ClCallEvent, data : any) => any)|null = null;
+	public onPrepareRequest : ((ev: ClCallEvent, data : EventMethodCallData) => any)|null = null;
 
-	public onLoading : ((data : any) => any)|null = null;
+	public onLoading : ((data : EventMethodCallData) => any)|null = null;
 
 	public onLoaded : ((handle : any, data : any)=>void)|null = null;
 
-	public onError : ((handle : any, error : any)=>void)|null = null;
+	public onError : ((handle : any, ev: EventError)=>void)|null = null;
+
+    public dataFetcher: (input: RequestInfo | URL, init: RequestInit, ev: ClCallEvent) => Promise<DataFetcherResponse> = this.defaultDataFetcher;
+
+    public async defaultDataFetcher(input: RequestInfo | URL, init: RequestInit, ev: ClCallEvent): Promise<DataFetcherResponse>
+    {
+        const fetchRes = await fetch(input, init);
+		const contentType = fetchRes.headers.get("Content-Type") ?? 'text/plain';
+        const isBlob = contentType === "octed/stream";
+
+        const res: DataFetcherResponse = {
+            isBlob,
+            response: isBlob
+				? await fetchRes.blob()
+				: (contentType === 'application/json' ? await fetchRes.json() : await fetchRes.text()),
+			contentType,
+			status: fetchRes.status,
+			statusText: fetchRes.statusText,
+			ok: fetchRes.ok,
+        }
+		//console.log('fetch res', res);
+		return res;
+    }
 
 	// @ts-ignore
 	protected async callMethod(methodName : string, args : any/* : IArguments*/, callOpts : CallOpts = { rawResult: false }, newDataType: new(data:any)=>any = null) : Promise<any> {
@@ -26,7 +61,7 @@ export class BaseCL {
 				newArgs.push(c);
 				c++;
 				uploadArgs.push(i);
-			} else if (arg instanceof FileList) {
+			} else if (typeof FileList !== 'undefined' && arg instanceof FileList) {
 				const newArg: any[] = [];
 				// @ts-ignore
 				for (const file of arg) {
@@ -42,7 +77,7 @@ export class BaseCL {
 			i++;
 		}
 
-		const srcData = {
+		const srcData: EventMethodCallData = {
 			name: methodName,
 			context: this.context,
 			pars: newArgs,
@@ -56,6 +91,7 @@ export class BaseCL {
 			postData = new FormData();
 			postData.append('request', JSON.stringify(srcData));
 			for (const k in uploads) {
+				// @ts-ignore
 				postData.append(k, uploads[k]);
 			}
 		} else {
@@ -78,21 +114,24 @@ export class BaseCL {
 		let response : any = undefined;
 
 		if (this.onLoading) loadingHandle = this.onLoading(srcData);
+		const eventError: EventError = {
+			methodArgs: args,
+			rawResponse: null,
+			error: null,
+		};
 
 		try {
-			const fetchRes = await fetch(ev.url, ev.requestInit);
-
-			if (fetchRes.headers["Content-Type"] === "octed/stream") {
-				response = await fetchRes.blob();
-				return response;
+			const fetchRes = await this.dataFetcher(ev.url, ev.requestInit, ev);
+			//console.log('fetchRes', fetchRes);
+			if (fetchRes.isBlob) {
+				return fetchRes.response;
 			}
 			else
 			{
-				response = await fetchRes.json();
+				response = fetchRes.response;
 
 				if (response.status == "ok")
 				{
-					// if (this.onLoaded) this.onLoaded(loadingHandle, response);
 					if (newDataType) response.response = new newDataType(response.response);
 					return response.response;
 				}
@@ -101,9 +140,9 @@ export class BaseCL {
 					throw new Error(response.exception);
 				}
 			}
-		} catch (exc) {
-			// if (this.onLoaded) this.onLoaded(loadingHandle, response);
-			if (this.onError) this.onError(loadingHandle, exc);
+		} catch (exc: any) {
+			eventError.error = exc;
+			if (this.onError) this.onError(loadingHandle, eventError);
 			throw exc;
 		} finally {
 			if (this.onLoaded) this.onLoaded(loadingHandle, response);
@@ -113,15 +152,6 @@ export class BaseCL {
 	constructor(url: string = "") {
 		this.url = url;
 	}
-
-
-	/* static getInstance(): self
-	{
-		if (!this.instance) {
-			this.instance = new this();
-		}
-		return this.instance;
-	} */
 }
 
 export type CallMethodResponse = {
@@ -140,4 +170,13 @@ export type CallOpts = {
 export type ClCallEvent = {
 	url: string,
 	requestInit: RequestInit,
+}
+
+export type DataFetcherResponse = {
+	ok: boolean,
+    response : any,
+    isBlob?: boolean,
+	status?: number,
+	statusText?: string,
+	contentType?: string,
 }