+++ src/TsLinkPhp/Bridges/Nette/TLApplication.php
@@ -10,6 +10,7 @@ use Nette\Application\Responses\FileResponse;
 use Nette\Application\Responses\TextResponse;
 use Nette\Http\IRequest;
 use Nette\Http\IResponse;
+use Nette\Utils\Json;
 use Nette\Utils\Strings;
 use Tracy\Debugger;
 
@@ -35,18 +36,44 @@ class TLApplication
 
     public array $middlewares = [];
 
+    public string $debugger = self::Debugger_Nette;
+
     public function addMiddleware(MiddlewareInterface $middleware) : void {
         $this->middlewares[] = $middleware;
     }
 
+    public const string Debugger_Nette = 'Nette';
+    public const string Debugger_Json = 'Json';
+    public const string Debugger_Text = 'Text';
+    public const string Debugger_Hide = 'Hide';
+
     /** @throws */
     public function run(): void
     {
-        set_error_handler([Debugger::class, 'errorHandler']);
+        if ($this->debugger === self::Debugger_Nette) set_error_handler([Debugger::class, 'errorHandler']);
+        $origin = $this->httpRequest->getHeader('origin');
+        if ($this->cors === true || (is_array($this->cors) && in_array($origin, $this->cors))) {
+            $requestMethod = $this->httpRequest->getMethod();
+            $allowedMethods = 'GET, POST, PUT, DELETE, OPTIONS';
+            $allowedHeaders = 'Content-Type, Authorization, X-Requested-With';
+            $maxAge = 1;
+
+            $this->httpResponse->setHeader('Access-Control-Allow-Origin', $origin);
+            $this->httpResponse->setHeader('Access-Control-Allow-Methods', $allowedMethods);
+            $this->httpResponse->setHeader('Access-Control-Allow-Headers', $allowedHeaders);
+            $this->httpResponse->setHeader('Access-Control-Allow-Credentials', 'true');
+            $this->httpResponse->setHeader('Access-Control-Max-Age', (string)$maxAge);
+
+            if ($requestMethod === 'OPTIONS' && $origin) {
+                $this->httpResponse->setCode(204);
+                return;
+            }
+
+        }
         try {
             $name = substr($this->httpRequest->getUrl()->path, strlen($this->urlPrefix));
             if ($name === '@code-gen' || $name === '@code-dump') {
-                if (!$this->codeGenFile && $name === '@code-gen') throw new BadRequestException('Not allowed', 405);
+                if (!$this->codeGenEnabled || (!$this->codeGenFile && $name === '@code-gen')) throw new BadRequestException('Not allowed', 405);
                 switch ($name) {
                     case '@code-gen':
                         file_put_contents($this->codeGenFile, $this->generateClientCode());
@@ -82,7 +109,26 @@ class TLApplication
                 }
             }
         } catch (\Throwable $e) {
-            Debugger::exceptionHandler($e);
+            if ($this->debugger === self::Debugger_Nette) Debugger::exceptionHandler($e);
+            else {
+                $this->httpResponse->setCode(IResponse::S500_InternalServerError);
+                $errorResponse = '';
+                switch($this->debugger) {
+                    case self::Debugger_Json:
+                        $errorResponse = Json::encode([
+                            'code' => $e->getCode(),
+                            'message' => $e->getMessage(),
+                            'file' => $e->getFile(),
+                            'trace' => $e->getTrace(),
+                        ]);
+                        break;
+                    case self::Debugger_Text:
+                        $errorResponse = (string)$e;
+                        break;
+                }
+                $this->httpResponse->setContentType($this->debugger === self::Debugger_Json ? 'application/json' : 'text/plain', 'utf-8');
+                (new TextResponse($errorResponse))->send($this->httpRequest, $this->httpResponse);
+            }
         }
     }
 
@@ -90,6 +136,7 @@ class TLApplication
         public string $urlPrefix,
         protected IRequest $httpRequest,
         protected IResponse $httpResponse,
+        protected array|bool $cors = false
     )
     {
     }